pyproject.toml
[project]
name = "termini"
version = "0.1.0"
description = ""
authors = [
    {name = "gevla"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai (>=1.28.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    {include ="termini"}
]
